+++ contracts/zkTrueUp/account/IAccountFacet.sol
@@ -30,7 +30,7 @@ interface IAccountFacet {
     /// @notice Withdraw Ether or ERC20 from ZkTrueUp
     /// @param token The token to be withdrawn
     /// @param amount The amount of the token to be withdrawn
-    function withdraw(IERC20 token, uint128 amount) external;
+    function withdraw(IERC20 token, uint256 amount) external;
 
     /// @notice Force withdraw Ether or ERC20 from ZkTrueUp
     /// @notice When the L2 system is down or user's asset is censored, user can do forceWithdraw to withdraw asset from ZkTrueUp
